extract auto reply stuff into ChatQuestionCarouselAutoReply
